some improved, but still buggy incremental indexing code

+++ plugin/jreframeworker/src/jreframeworker/core/JReFrameworkerProject.java
@@ -13,16 +13,23 @@ import java.util.Set;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.transform.TransformerException;
 
+import org.apache.commons.io.FileDeleteStrategy;
+import org.eclipse.core.resources.ICommand;
 import org.eclipse.core.resources.IProject;
+import org.eclipse.core.resources.IProjectDescription;
 import org.eclipse.core.resources.IResource;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.NullProgressMonitor;
+import org.eclipse.core.runtime.Path;
 import org.eclipse.jdt.core.IClasspathEntry;
 import org.eclipse.jdt.core.IJavaProject;
 import org.eclipse.jdt.core.IPackageFragmentRoot;
 import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.jdt.internal.core.ClasspathEntry;
 import org.xml.sax.SAXException;
+
+import jreframeworker.builder.JReFrameworkerBuilder;
+import jreframeworker.log.Log;
 
 @SuppressWarnings("restriction")
 public class JReFrameworkerProject {
@@ -35,12 +42,84 @@ public class JReFrameworkerProject {
 		this.jProject = JavaCore.create(project);
 	}
 	
+	public BuildFile getBuildFile(){
+		return BuildFile.getOrCreateBuildFile(jProject);
+	}
+	
 	/**
 	 * Returns the Eclipse project resource
 	 * @return
 	 */
 	public IProject getProject(){
 		return project;
+	}
+	
+	/**
+	 * Returns the Eclipse project resource
+	 * @return
+	 */
+	public IJavaProject getJavaProject(){
+		return jProject;
+	}
+	
+	public void clean() throws CoreException {
+		try {
+			File buildDirectory = project.getFolder(JReFrameworker.BUILD_DIRECTORY).getLocation().toFile();
+
+			// restore the classpath
+			restoreOriginalClasspathEntries();
+
+			if (buildDirectory.exists()) {
+				clearProjectBuildDirectory(buildDirectory);
+			}
+		} catch (Exception e) {
+			Log.error("Error cleaning " + project.getName(), e);
+		}
+		project.refreshLocal(IResource.DEPTH_INFINITE, new NullProgressMonitor());
+	}
+
+	private void clearProjectBuildDirectory(File buildDirectory) throws IOException {
+		for(File file : buildDirectory.listFiles()){
+			if(file.isDirectory()){
+				File directory = file;
+				clearProjectBuildDirectory(directory);
+				directory.delete();
+			} else {
+				FileDeleteStrategy.FORCE.delete(file);
+			}
+		}
+	}
+	
+	public void removeJavaNature() throws CoreException {
+		IProjectDescription description = getProject().getDescription();
+		ICommand[] commands = description.getBuildSpec();
+		for (int i = 0; i < commands.length; ++i) {
+			if (commands[i].getBuilderName().equals(JavaCore.BUILDER_ID)) {
+				ICommand[] newCommands = new ICommand[commands.length - 1];
+				System.arraycopy(commands, 0, newCommands, 0, i);
+				System.arraycopy(commands, i + 1, newCommands, i, commands.length - i - 1);
+				description.setBuildSpec(newCommands);
+				project.setDescription(description, null);			
+				return;
+			}
+		}
+	}
+	
+	public void addJavaNature() throws CoreException {
+		IProjectDescription desc = project.getDescription();
+		ICommand[] commands = desc.getBuildSpec();
+		for (int i = 0; i < commands.length; ++i) {
+			if (commands[i].getBuilderName().equals(JavaCore.BUILDER_ID)) {
+				return;
+			}
+		}
+		ICommand[] newCommands = new ICommand[commands.length + 1];
+		System.arraycopy(commands, 0, newCommands, 0, commands.length);
+		ICommand command = desc.newCommand();
+		command.setBuilderName(JReFrameworkerBuilder.BUILDER_ID);
+		newCommands[newCommands.length - 1] = command;
+		desc.setBuildSpec(newCommands);
+		project.setDescription(desc, null);
 	}
 	
 	/**
@@ -51,7 +130,7 @@ public class JReFrameworkerProject {
 	 * @throws ParserConfigurationException
 	 */
 	public Set<String> listTargets() throws SAXException, IOException, ParserConfigurationException {
-		return BuildFile.getOrCreateBuildFile(project).getTargets();
+		return getBuildFile().getTargets();
 	}
 	
 	/**
@@ -60,8 +139,12 @@ public class JReFrameworkerProject {
 	 * @throws URISyntaxException 
 	 */
 	public void addTarget(File targetLibrary) throws TransformerException, ParserConfigurationException, SAXException, IOException, URISyntaxException, CoreException {
-		addProjectLibrary(jProject, targetLibrary);
-		BuildFile.getOrCreateBuildFile(project).addTarget(targetLibrary.getName());
+		String entry = addProjectLibrary(jProject, targetLibrary);
+		
+		// update the build file
+		BuildFile buildFile = getBuildFile();
+		buildFile.addTarget(targetLibrary.getName());
+		buildFile.addOriginalClasspathEntry(entry);
 	}
 	
 	/**
@@ -76,15 +159,19 @@ public class JReFrameworkerProject {
 	 * @throws CoreException
 	 */
 	public void addTarget(File targetLibrary, String relativeLibraryDirectory) throws TransformerException, ParserConfigurationException, SAXException, IOException, URISyntaxException, CoreException {
-		addProjectLibrary(jProject, targetLibrary, relativeLibraryDirectory);
-		BuildFile.getOrCreateBuildFile(project).addTarget(targetLibrary.getName());
+		String entry = addProjectLibrary(jProject, targetLibrary, relativeLibraryDirectory);
+		
+		// update the build file
+		BuildFile buildFile = getBuildFile();
+		buildFile.addTarget(targetLibrary.getName());
+		buildFile.addOriginalClasspathEntry(entry);
 	}
 	
 	/**
 	 * Removes a target from the JReFrameworker project
 	 */
 	public void removeTarget(String target) throws TransformerException, ParserConfigurationException, SAXException, IOException {
-		BuildFile.getOrCreateBuildFile(project).removeTarget(target);
+		BuildFile.getOrCreateBuildFile(jProject).removeTarget(target);
 	}
 	
 	public void refresh() throws CoreException {
@@ -100,8 +187,85 @@ public class JReFrameworkerProject {
 	 * @throws MalformedURLException
 	 * @throws CoreException
 	 */
-	private static void addProjectLibrary(IJavaProject jProject, File library) throws IOException, URISyntaxException, MalformedURLException, CoreException {
-		addProjectLibrary(jProject, library, null);
+	private static String addProjectLibrary(IJavaProject jProject, File library) throws IOException, URISyntaxException, MalformedURLException, CoreException {
+		return addProjectLibrary(jProject, library, null);
+	}
+	
+	/**
+	 * Replaces the classpath jar entry with the given jar
+	 * @param jarName
+	 * @param updatedLibrary
+	 * @throws IOException 
+	 * @throws CoreException 
+	 */
+	public void updateProjectLibrary(String jarName, File updatedLibrary) throws IOException, CoreException {
+		updatedLibrary = updatedLibrary.getCanonicalFile();
+		String updatedLibraryPath = updatedLibrary.getCanonicalPath();
+		File projectRoot = project.getLocation().toFile().getCanonicalFile();
+		
+		boolean isUpdatedLibraryContainedInProject = false;
+		File parent = updatedLibrary.getParentFile();
+		while(parent != null){
+			if(parent.equals(projectRoot)){
+				isUpdatedLibraryContainedInProject = true;
+				break;
+			} else {
+				parent = parent.getParentFile();
+			}
+		}
+		
+		// if the updated library is inside the project, then make the path relative
+		// otherwise we must use the absolution path
+		if(isUpdatedLibraryContainedInProject){
+			String base = projectRoot.getCanonicalPath();
+			String relativeFilePath = updatedLibrary.getCanonicalPath().substring(base.length());
+			if(relativeFilePath.charAt(0) == File.separatorChar){
+				relativeFilePath = relativeFilePath.substring(1);
+			}
+			updatedLibraryPath = relativeFilePath;
+		}
+		
+		
+		// create a classpath entry for the library
+		IClasspathEntry updatedLibraryEntry;
+		if(isUpdatedLibraryContainedInProject){
+			updatedLibraryPath = updatedLibraryPath.replace(File.separator, "/");
+	    	// library is at some path relative to project root
+	    	updatedLibraryEntry = new org.eclipse.jdt.internal.core.ClasspathEntry(
+	    	        IPackageFragmentRoot.K_BINARY,
+	    	        IClasspathEntry.CPE_LIBRARY, jProject.getProject().getFile(updatedLibraryPath).getLocation(),
+	    	        ClasspathEntry.INCLUDE_ALL, // inclusion patterns
+	    	        ClasspathEntry.EXCLUDE_NONE, // exclusion patterns
+	    	        null, null, null, // specific output folder
+	    	        false, // exported
+	    	        ClasspathEntry.NO_ACCESS_RULES, false, // no access rules to combine
+	    	        ClasspathEntry.NO_EXTRA_ATTRIBUTES);
+	    } else {
+	    	// library is outside the project, using absolute path
+	    	updatedLibraryEntry = new org.eclipse.jdt.internal.core.ClasspathEntry(
+	    	        IPackageFragmentRoot.K_BINARY,
+	    	        IClasspathEntry.CPE_LIBRARY, new Path(updatedLibraryPath),
+	    	        ClasspathEntry.INCLUDE_ALL, // inclusion patterns
+	    	        ClasspathEntry.EXCLUDE_NONE, // exclusion patterns
+	    	        null, null, null, // specific output folder
+	    	        false, // exported
+	    	        ClasspathEntry.NO_ACCESS_RULES, false, // no access rules to combine
+	    	        ClasspathEntry.NO_EXTRA_ATTRIBUTES);
+	    }
+		
+		// search through the classpath's existing entries and replace the corresponding library entry
+	    IClasspathEntry[] entries = jProject.getRawClasspath();
+	    for(int i=0; i< entries.length; i++){
+	    	if(entries[i].getPath().toFile().getName().equals(jarName)){
+	    		entries[i] =  updatedLibraryEntry;
+	    		// assuming there is only one library with the same name...
+	    		break;
+	    	}
+	    }
+	    jProject.setRawClasspath(entries, null);
+	    
+	    // refresh project
+	 	jProject.getProject().refreshLocal(IResource.DEPTH_INFINITE, new NullProgressMonitor());
 	}
 	
 	/**
@@ -114,8 +278,16 @@ public class JReFrameworkerProject {
 	 * @throws MalformedURLException
 	 * @throws CoreException
 	 */
-	private static void addProjectLibrary(IJavaProject jProject, File libraryToAdd, String relativeDirectoryPath) throws IOException, URISyntaxException, MalformedURLException, CoreException {
-	    // copy the jar file into the project
+	private static String addProjectLibrary(IJavaProject jProject, File libraryToAdd, String relativeDirectoryPath) throws IOException, URISyntaxException, MalformedURLException, CoreException {
+		
+		// only add the project library to the classpath if its not already there
+		for(IClasspathEntry entry : jProject.getRawClasspath()){
+			if(entry.getPath().toFile().getName().endsWith(libraryToAdd.getName())){
+				return entry.getPath().toString();
+			}
+		}
+		
+	    // copy the jar file into the project (if its not already there)
 	    InputStream libraryInputStream = new BufferedInputStream(new FileInputStream(libraryToAdd));
 	    File libDirectory;
 	    if(relativeDirectoryPath == null || relativeDirectoryPath.equals("")){
@@ -126,7 +298,9 @@ public class JReFrameworkerProject {
 	    }
 		libDirectory.mkdirs();
 		File library = new File(libDirectory.getCanonicalPath() + File.separatorChar + libraryToAdd.getName());
-		Files.copy(libraryInputStream, library.toPath());
+		if(!library.exists()){
+			Files.copy(libraryInputStream, library.toPath());
+		}
 
 		// refresh project
 		jProject.getProject().refreshLocal(IResource.DEPTH_INFINITE, new NullProgressMonitor());
@@ -167,6 +341,29 @@ public class JReFrameworkerProject {
 	    
 	    // refresh project
 	 	jProject.getProject().refreshLocal(IResource.DEPTH_INFINITE, new NullProgressMonitor());
+	 	
+	 	return relativeLibraryEntry.getPath().toString();
+	}
+
+	/**
+	 * Restores the original classpath entries of the project
+	 * @throws SAXException
+	 * @throws IOException
+	 * @throws ParserConfigurationException
+	 * @throws CoreException
+	 */
+	public void restoreOriginalClasspathEntries() throws SAXException, IOException, ParserConfigurationException, CoreException {
+		for(String entry : getBuildFile().getOriginalClasspathEntries()){
+			File library = new File(entry);
+			if(library.exists()){
+				updateProjectLibrary(library.getName(), library);
+			} else {
+				library = project.getFile(entry).getLocation().toFile();
+				if(library.exists()){
+					updateProjectLibrary(library.getName(), library);
+				}
+			}
+		}
 	}
 
 }
